Update references to utility library for FFT calculation
