+++ config.xml
@@ -93,7 +93,7 @@
 			<PosY>23</PosY>
 			<PosX>0</PosX>
 			<Width>1920</Width>
-			<Heigth>977</Heigth>
+			<Heigth>979</Heigth>
 		</Frame>
 		<UserDir>
 			<MSEED>/Users/mth/mth/Projects/xmax</MSEED>
