+++ settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'XMAX'
+
+include 'asl-java-utils'
